+++ mgkit/workflow/filter_gff.py
@@ -144,7 +144,7 @@ import logging
 import functools
 import click
 import pandas
-
+from tqdm import tqdm
 import mgkit
 from . import utils
 from ..io import gff, fasta
@@ -221,10 +221,12 @@ def find_comparison(comparison):
 @click.option('-c', '--comparison', default='ge', show_default=True,
               type=click.Choice(['gt', 'ge', 'lt', 'le']),
               help='Type of comparison (e.g. ge -> greater than or equal to)')
+@click.option('--progress', default=False, is_flag=True,
+              help="Shows Progress Bar")
 @click.argument('input-file', type=click.File('rb'), default='-')
 @click.argument('output-file', type=click.File('wb'), default='-')
 def filter_perseq(verbose, sorted, attribute, function, value, comparison,
-                  input_file, output_file):
+                  progress, input_file, output_file):
 
     mgkit.logger.config_log(level=logging.DEBUG if verbose else logging.INFO)
 
@@ -249,6 +251,9 @@ def filter_perseq(verbose, sorted, attribute, function, value, comparison,
         grouped = viewvalues(gff.group_annotations(
             file_iterator, lambda x: x.seq_id
         ))
+
+    if progress:
+        grouped = tqdm(grouped)
 
     for annotations in grouped:
         threshold = perseq_calc_threshold(
@@ -275,10 +280,12 @@ def filter_perseq(verbose, sorted, attribute, function, value, comparison,
               help='Minimum coverage for the contig/strand')
 @click.option('-r', '--rename', default=False, is_flag=True,
               help='Emulates BLAST in reading the FASTA file (keeps only the header before the first space)')
+@click.option('--progress', default=False, is_flag=True,
+              help="Shows Progress Bar")
 @click.argument('input-file', type=click.File('rb'), default='-')
 @click.argument('output-file', type=click.File('wb'), default='-')
 def coverage_command(verbose, reference, strand_specific, sorted, min_coverage,
-                     rename, input_file, output_file):
+                     rename, progress, input_file, output_file):
     mgkit.logger.config_log(level=logging.DEBUG if verbose else logging.INFO)
     LOG.info(
         'Writing to file (%s)',
@@ -308,6 +315,9 @@ def coverage_command(verbose, reference, strand_specific, sorted, min_coverage,
         seq_id: len(seq)
         for seq_id, seq in seq_iter
     }
+
+    if progress:
+        grouped = tqdm(grouped)
 
     for annotations in grouped:
         covered = ranges_length(
@@ -435,10 +445,12 @@ def validate_params(ctx, param, values, convert=str):
 @click.option('--num-lt', multiple=True,
               callback=functools.partial(validate_params, convert=float),
               help="Same as '--num-ge' but 'value' is a number which is less than")
+@click.option('--progress', default=False, is_flag=True,
+              help="Shows Progress Bar")
 @click.argument('input-file', type=click.File('rb'), default='-')
 @click.argument('output-file', type=click.File('wb'), default='-')
 def filter_values(verbose, str_eq, str_in, num_eq, num_ge, num_le, num_gt, num_lt,
-                  input_file, output_file):
+                  progress, input_file, output_file):
     mgkit.logger.config_log(level=logging.DEBUG if verbose else logging.INFO)
 
     LOG.info(
@@ -448,7 +460,11 @@ def filter_values(verbose, str_eq, str_in, num_eq, num_ge, num_le, num_gt, num_l
 
     filters = setup_filters(str_eq, str_in, num_eq, num_ge, num_le, num_gt, num_lt)
 
-    for annotation in gff.parse_gff(input_file, gff_type=gff.from_gff):
+    iterator = gff.parse_gff(input_file, gff_type=gff.from_gff)
+    if progress:
+        iterator = tqdm(iterator)
+
+    for annotation in iterator:
         if all(filter_func(annotation) for filter_func in filters):
             annotation.to_file(output_file)
 
@@ -491,10 +507,12 @@ def make_choose_func(values):
 @click.option('-a', '--sort-attr', default='bitscore', show_default=True,
               type=click.Choice(['bitscore', 'identity', 'length']),
               help='Attribute to sort annotations before filtering (default bitscore)')
+@click.option('--progress', default=False, is_flag=True,
+              help="Shows Progress Bar")
 @click.argument('input-file', type=click.File('rb'), default='-')
 @click.argument('output-file', type=click.File('wb'), default='-')
-def filter_overlaps(verbose, size, sorted, choose_func, sort_attr, input_file,
-                    output_file):
+def filter_overlaps(verbose, size, sorted, choose_func, sort_attr, progress,
+                    input_file, output_file):
 
     mgkit.logger.config_log(level=logging.DEBUG if verbose else logging.INFO)
 
@@ -518,6 +536,9 @@ def filter_overlaps(verbose, size, sorted, choose_func, sort_attr, input_file,
         overlap=size,
         choose_func=choose_func
     )
+
+    if progress:
+        grouped = tqdm(grouped)
 
     for annotations in grouped:
         filtered = filter_gff.filter_annotations(